Changed the names of the examples
