+++ examples/vector_add/main.c
@@ -13,76 +13,62 @@
 #include <otto/status.h>
 #include <otto/vector.h>
 
-#define MAX_SOURCE_SIZE (0x100000)
-
 int main(void) {
-  log_info("Creating the parameters");
+  /* Runtime creation */
+  log_info("Creating the runtime");
+  otto_runtime_t ctx;
+  otto_kernelht_t *ht = NULL;
+  otto_runtime_new(NULL, NULL, OTTO_DEVICE_GPU, ht, &ctx);
+  otto_runtime_load_kernels(&ctx, OTTO_KERNELS_CORE, "");
+
+  /* Create the vectors for the kernel call */
+  log_info("Creating the vectors");
   otto_vector_t a;
   otto_vector_t b;
   otto_vector_t out;
   const int LIST_SIZE = 256;
-  OTTO_CALL(otto_vector_with_capacity(LIST_SIZE, sizeof(int), &a),
-            "Failed allocating A");
-  OTTO_CALL(otto_vector_with_capacity(LIST_SIZE, sizeof(int), &b),
-            "Failed allocating B");
-  OTTO_CALL(otto_vector_with_capacity(LIST_SIZE, sizeof(int), &out),
-            "Failed allocating OUT");
+  otto_vector_with_capacity(LIST_SIZE, sizeof(int), &a);
+  otto_vector_with_capacity(LIST_SIZE, sizeof(int), &b);
+  otto_vector_with_capacity(LIST_SIZE, sizeof(int), &out);
   for (int i = 0; i < LIST_SIZE; i++) {
     otto_vector_push(&a, &i);
     int v = LIST_SIZE - i;
     otto_vector_push(&b, &v);
   }
-
-  log_info("Creating the runtime");
-  otto_runtime_t ctx;
-  otto_kernelht_t *ht = NULL;
-  OTTO_CALL(otto_runtime_new(NULL, NULL, OTTO_DEVICE_GPU, ht, &ctx),
-            "Could not initialize the runtime");
+  otto_vector_setwrite(&out);
 
   log_info("Creating the buffers in device memory");
-  OTTO_CALL(otto_vector_todevice_read(&a, &ctx), "Failed registering A");
-  OTTO_CALL(otto_vector_todevice_read(&b, &ctx), "Failed registering B");
-  OTTO_CALL(otto_vector_todevice_write(&out, &ctx), "Failed registering OUT");
+  otto_vector_todevice(&a, &ctx);
+  otto_vector_todevice(&b, &ctx);
+  otto_vector_todevice(&out, &ctx);
 
-  log_info("Creating the program");
-  const char *files[] = {OTTO_CLKERNEL("vector/elementary.cl")};
-  otto_program_t prog;
-  OTTO_CALL(otto_program_from_files(&ctx, files, 1, "", &prog),
-            "Failed to create the program");
-
-  log_info("Creating the kernel");
-  OTTO_CALL(otto_kernel_new(&prog, "otto_vector_add", 3, &ctx, NULL),
-            "Failed creating kernel");
-
+  /* Call the kernel */
   log_info("Creating hparams");
   otto_kernel_args_t hparams = {
       .work_dim = 1,
       .global_size = out.capacity,
       .local_size = 64,
   };
-
   log_info("Calling kernel directly from runtime");
-  OTTO_CALL(otto_runtime_call_kernel_binop(&ctx, "otto_vector_add__i32", &hparams,
-                                           &a, &b, &out),
-            "Failed calling kernel");
-
+  otto_runtime_call_kernel_binop(&ctx, "otto_vector_add__i32", &hparams, &a, &b,
+                                 &out);
   log_info("Reading from the output buffer");
-  OTTO_CALL(otto_vector_tohost(&out, 0), "Failed fetching OUT to host");
+  otto_vector_tohost(&out, 0);
 
   log_info("Displaying results");
   for (int i = 0; i < LIST_SIZE; i++) {
     int a_v, b_v, out_v;
-    OTTO_CALL(otto_vector_get(&a, i, &a_v), "Could not get A[%d]", i);
-    OTTO_CALL(otto_vector_get(&b, i, &b_v), "Could not get B[%d]", i);
-    OTTO_CALL(otto_vector_get(&out, i, &out_v), "Could not get OUT[%d]", i);
+    otto_vector_get(&a, i, &a_v);
+    otto_vector_get(&b, i, &b_v);
+    otto_vector_get(&out, i, &out_v);
     printf("%d + %d = %d\n", a_v, b_v, out_v);
   }
 
+  /* Clean the program */
   log_info("Doing final cleanup");
-  OTTO_CALL(otto_vector_cleanup(&a), "Failed cleaning A");
-  OTTO_CALL(otto_vector_cleanup(&b), "Failed cleaning B");
-  OTTO_CALL(otto_vector_cleanup(&out), "Failed cleaning OUT");
-  OTTO_CALL(otto_runtime_cleanup(&ctx), "Failed cleaning runtime");
-  OTTO_CALL(otto_program_cleanup(&prog), "Failed cleaning program");
+  otto_vector_cleanup(&a);
+  otto_vector_cleanup(&b);
+  otto_vector_cleanup(&out);
+  otto_runtime_cleanup(&ctx);
   return 0;
 }
